Improvements in file organization
