+++ tokens/figma/utils/constants.mjs
@@ -1,0 +1,20 @@
+export const VARIABLE_TYPES = {
+  COLOR: "COLOR",
+  FLOAT: "FLOAT",
+  STRING: "STRING",
+  FONT_WEIGHT: "FONT_WEIGHT",
+  FONT_SIZE: "FONT_SIZE",
+  LINE_HEIGHT: "LINE_HEIGHT",
+};
+
+export const COLLECTION_NAMES = {
+  BRAND: "Brand",
+  COLOR_SCHEME: "Mode",
+  PALETTE: "Palette",
+};
+
+export const MODE_NAMES = {
+  DEFAULT: "Mode 1",
+  LIGHT: "Light",
+  DARK: "Dark",
+};
